refactor stack organization; init lambdaApi stack
