+++ lib/index.ts
@@ -1,5 +1,6 @@
-import * as sst from "@serverless-stack/resources";
+import { App } from "@serverless-stack/resources";
+import SavourDashboardStack from "./stacks/SavourDashboardStack";
 
-export default function main(app: sst.App): void {
-
+export default function main(app: App): void {
+	new SavourDashboardStack(app, "savour-dashboard");
 }
